Initial separation of MoveIt planner from RosClient
