+++ src/compas_fab/backends/ros/__init__.py
@@ -1,23 +1,26 @@
 """
-********************************************************************************
+*******************************************************************************
 compas_fab.backends.ros
-********************************************************************************
+*******************************************************************************
 
 .. module:: compas_fab.backends.ros
 
-Package with functionality to interact with the `ROS platform <http://ros.org/>`_.
+Package with functionality to interact with `ROS <http://ros.org/>`_.
 
 .. autosummary::
     :toctree: generated/
 
     RosClient
     RosError
+    RosValidationError
 
 """
 
 from __future__ import absolute_import
 
-from .client import RosClient
-from .client import RosError
-from .direct_ur_action_client import DirectUrActionClient
+from .client import *
+from .exceptions import *
+from .direct_ur_action_client import *
 from .messages import *
+
+__all__ = [name for name in dir() if not name.startswith('_')]
